fix: clean 없애니까 앱 안 죽음
