+++ app/src/main/java/com/eddy/nrf/presentation/ui/BikeViewModel.kt
@@ -13,11 +13,11 @@ class BikeViewModel : ViewModel() {
     private val _uiState = MutableStateFlow(BikeUiState())
     val uiState: StateFlow<BikeUiState> = _uiState.asStateFlow()
 
-    init {
-        updateUi(1F)
-    }
+//    init {
+//        onBluetoothDataReceived(1F)
+//    }
 
-    fun updateUi(gear: Float) {
+    fun onBluetoothDataReceived(gear: Float) {
         viewModelScope.launch {
             _uiState.update {
                 it.copy(
